requirements.txt
fastapi["standard"]
uvicorn
mysql-connector-python
python-dotenv
